split up runtime suite
