+++ test/runtime/runtime_1.test.js
@@ -1,0 +1,5 @@
+// @vitest-environment jsdom
+
+import { run_shard } from './runtime.shared.js';
+
+run_shard(1, 2);
